enforces types for component input values, replaces anchor tags with Link component, tweaks some link padding; tweaks some link coloring; introduces SocialLink component
