+++ scss/_variables.scss
@@ -1,5 +1,5 @@
 // link hover colors and footer text color
-$highlight2-color: #ff8c42;
+$highlight2-color: #ff8800;
 $highlight3-color: #cb769e;
 
 $link-color: rgba(0, 0, 139, 0.8);
