+++ src/components/Courses.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react';
 import ICourse from '../interfaces/ICourse';
 import GetDisciplineByDisciplineType from '../functions/GetDisciplineByDisciplineType';
+import Link from './Link';
 
 const Courses: FC<{ items: ICourse[] }> = function({ items }) {
   const disciplineType = GetDisciplineByDisciplineType(items[0].discipline);
@@ -12,11 +13,11 @@ const Courses: FC<{ items: ICourse[] }> = function({ items }) {
         {items.sort((a: ICourse, b: ICourse) => a.discipline - b.discipline || (new Date(a.completionDate).getTime() - new Date(b.completionDate).getTime()))
           .map((c: ICourse) => {
             const tooltipForCourse = `${c.provider} course completed on ${new Date(c.completionDate).toDateString()}`;
+            const logoUrl = `./assets/${c.provider.toLowerCase()}.png`;
             return (
               <div className="text-bold mt-1" key={++cIdx}>
-                <a href={c.completionUrl} target="_blank" rel="noreferrer" title={tooltipForCourse}>
-                  {c.name}
-                </a>
+                <img src={logoUrl} alt="" height="24px" title={c.provider} />&nbsp;
+                <Link name={c.name} url={c.completionUrl} tooltip={tooltipForCourse} />
               </div>
           );
         })}
