+++ src/components/Disciplines.tsx
@@ -5,6 +5,7 @@ import SimpleSkill from "./SimpleSkill";
 import GetDisciplineByDisciplineType from "../functions/GetDisciplineByDisciplineType";
 import Courses from "./Courses";
 import ICourse from "../interfaces/ICourse";
+import Link from "./Link";
 
 const Disciplines: FC<{ skills: ISkill[]; courses: ICourse[];}> = function ({ skills, courses }) {
   const distinctDisciplines = Array.from(new Set(skills.map((a) => a.discipline)));
@@ -17,9 +18,8 @@ const Disciplines: FC<{ skills: ISkill[]; courses: ICourse[];}> = function ({ sk
         and is one of the main reasons I gravitated to this career. It seems
         like the more we know, the more we know we don't know. The proficiency 
         values below align to the knowledge scale defined&nbsp;
-        <a href="https://hr.uiowa.edu/careers/competencies/proficiency-levels" target="_blank" rel="noreferrer">
-          here
-        </a>. I have a personal Pluralsight license that I use to keep up with any
+        <Link name="here" url="https://hr.uiowa.edu/careers/competencies/proficiency-levels" tooltip="" />.
+        I have a personal Pluralsight license that I use to keep up with any
         new technology I am exposed to. I'm my worst critic and tend to 
         underestimate my abilities while overdelivering results.
       </p>
