+++ src/toc.json
@@ -429,7 +429,7 @@
     },
     {
       "discipline": 0,
-      "provider": "Linked In",
+      "provider": "LinkedIn",
       "completionUrl": "https://www.linkedin.com/learning/certificates/067337ca88b8da7e780f41f97442840657f437ca3b9453bb5070b3f05fb34270?trk=backfilled_certificate",
       "name": "Node.js Essential Training (2019)",
       "completionDate": "May 09, 2021"
